Add flexible cut guides
This adds a few different styles of guide, additionally allowing for setting the color.

For now, this is a library-defined setting that default to the "crosshair" style, as this is the most flexible and ink-saving approach.

In the future, this could be a public-facing setting, allowing for specific components applying different guides/colors, depending on context. Sometimes one style might just work better than another.

Consequently, this commit also makes the existing guide-making functions obsolete.