Supress double/single-sided mix warning
I realized that there's actually no problem in mixing layouts, since a user can just choose to print a specific range of pages, rather than the whole thing.

So for now, i'm leaving this warning in as a notice, but really it might be best to just remove it entirely.