Replace `web` document with `proof`
Initially, the thought behind `web` was to produce something that could be distributed as a static-site or similar. However, in reality, this endpoint serves only to proof your layouts; i.e. previewing your components to find layout/design issues.

So with this commit, we abandon the idea of `web` and formally make it a proofing endpoint.