Rename `.individual` as `.png`
I'm not sure that we want to support other image format (bmp, jpeg, tiff are all potential formats we _could_ support), but if we do, these would probably come in the form of additional cases.

Alternatively, keeping something similar to `.individual`, but including a format specifier as a parameter is also a possibility. Not sure which is preferable. Either way, `.png` is better than `.individual` at this point, since it is immediately clear what the result will be.