Fix custom properties for backsides
Previously, it was not possible to ever apply any custom properties to the backside unless you created the component yourself (e.g. also redundantly specifying sizing and partitioning).

This commit fixes that issue by providing a closure that can be used intermittently to customize the backing component, giving you a fluent option.